fix: 🐛move useAppStore: useAppStore.ts -> store.ts
