+++ src/hocs/withUnAuthGuard.tsx
@@ -1,7 +1,7 @@
 import { useRouter } from 'next/router';
 import { useEffect } from 'react';
 
-import useAppStore from '@features/useAppStore';
+import { useAppStore } from '@features/store';
 
 export default function withUnAuthGuard<P>(
   AppComponent: React.ComponentType<P>,
